% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudocells.R
\name{calculate.pseudocells}
\alias{calculate.pseudocells}
\title{Calculates pseudocells from a Seurat object}
\usage{
calculate.pseudocells(
  seurat,
  seeds = 0.2,
  nn = 10,
  reduction = "pca",
  dims = 1:20,
  features = NULL,
  cells = NULL
)
}
\arguments{
\item{seurat}{The seurat object, with pre-computed PCA or other reductions, and the relevant clustering as IDs}

\item{seeds}{The proportion of cells to be used as seeds. Alternatively, a string with the name of the seeds to use. Numeric between 0.1 and 0.9 or string. Default 0.2}

\item{nn}{Number of nearest neighbors to compute and use for pseudocell aggregation. Default 10}

\item{reduction}{The name of the reduction to use. Should be present in the @reductions slot of the seurat object. Default is "pca"}

\item{dims}{The relevant dimensions that will be used to compute nearest neighbors. Default 1:20}

\item{features}{The features to be used. Takes a string of feature names as present in the expression matrices. Defaults to NULL, which will use all the genes.}

\item{cells}{The clusters or identities to be used. Takes a string of identities as present in the @active.ident slot. Defaults to NULL, which will use all the identities.}
}
\value{
A seurat object of aggregated pseudocells. With average expression. The slot misc contains the pseudocells dataframe, with each original cell and its assigned pseudocell, if no pseudocell is assigned then 00
}
\description{
This function calculates pseudocells from a Seurat object, based on pre-calculated cell clusters and dimentionality reduction. WARNING: This might be time consuming, depending on the size of the dataset.
}
\examples{
ps.pbmc_small=calculate.pseudocells(Seurat::pbmc_small, dims = 1:10)
}
