% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softthreshold.R
\name{choose.sfttp}
\alias{choose.sfttp}
\title{Chooses a soft-tresholding power}
\usage{
choose.sfttp(datexpr)
}
\arguments{
\item{datexpr}{A matrix of gene expression. This SHOULD BE the expression data normally stored under ´@assays$RNA@data´ for single cell objects or ´@assays$RNA@counts´ for pseudocells. This has to contain ONLY the genes used for analysis.}
}
\value{
A number representing the soft-thresholding power with the highest scale-free topology fit index. And a message about it
}
\description{
This function examines the data to find a suitable soft-thresholding power. It also helps to know if there is a scale-free topology.
}
\examples{
ps.pbmc_small=choose.sfttp(SeuratObject::pbmc_small@assays$RNA@data)
}
