% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{GatherData}
\alias{GatherData}
\alias{GatherData.Seurat}
\alias{GatherData.SingleCellExperiment}
\title{GatherData}
\usage{
GatherData(object, ...)

\method{GatherData}{Seurat}(object, assay, slot_use, ...)

\method{GatherData}{SingleCellExperiment}(object, assay, ...)
}
\arguments{
\item{object}{scRNA-seq data object to pull data from}

\item{...}{Additional arguments}

\item{assay}{Assay to pull data from}

\item{slot_use}{Slot to pull data from.  Ignored for SingleCellExperiment objects}
}
\value{
matrix

matrix
}
\description{
GatherData
}
