% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterWGCNA.R
\name{makeDendroandColors}
\alias{makeDendroandColors}
\title{Make Dendrogram and Colors}
\usage{
makeDendroandColors(dissTOM, min.module.size)
}
\arguments{
\item{min.module.size}{}
}
\value{

}
\description{
Make Dendrogram and Colors
}
