% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MmLimbE155.scWGCNA}
\alias{MmLimbE155.scWGCNA}
\title{Prices of 50,000 round cut diamonds.}
\format{
A scWGCNA list object, used as example
}
\source{
This package
}
\usage{
MmLimbE155.scWGCNA
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
