% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MmvGg.comparative}
\alias{MmvGg.comparative}
\title{A pre-computed comparative scWGCNA list calculated from the Mouse limb and Chicken data.}
\format{
A comparative scWGCNA list object, used as example
}
\source{
This package
}
\usage{
MmvGg.comparative
}
\description{
A comparative scWGCNA data list containing the different statistics and calculations
}
\keyword{datasets}
