% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my.small_GgLimbHH29}
\alias{my.small_GgLimbHH29}
\title{A pre-analyzed, sub-sampled Seurat object, of single cell RNA-seq from chicken Limbs at stage HH29}
\format{
A Seurat object, used as example
}
\source{
https://doi.org/10.1002/dvdy.384 sub-sampled for this package
}
\usage{
my.small_GgLimbHH29
}
\description{
A Seurat object containing different analyses and other data
}
\keyword{datasets}
