% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.WGCNA.dendro}
\alias{plot.WGCNA.dendro}
\title{Plots the gene / module WGCNA tree, or trees}
\usage{
\method{plot}{WGCNA.dendro}(scWGCNA.data, tree)
}
\arguments{
\item{scWGCNA.data}{scWGCNA.data. An scWGCNA.data object, as calculated by run.scWGCNA().}

\item{history}{Logial. Should all the trees in the scWGCNA iteration be returned?. Default is FALSE, only plots the last tree.}
}
\value{
If history is FALSE (default) plots the gene / module WGCNA. If history is TRUE, returns a list of all plots generated during the iterations.
}
\description{
This function will help us plot the gene / module WGCNA tree. If history is TRUE, it will return all trees as iterated by scWCGNA
}
\examples{
# Plot the WGCNA tree
plot.WGCNA.dendro(pbmc_small.scWGCNA)

# Plot the first WGCNA tree of the iteration
plot.WGCNA.dendro(pbmc_small.scWGCNA, 1)

}
