% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runscWGCNA.R
\name{run.scWGCNA}
\alias{run.scWGCNA}
\title{Runs a semi-automatic, iterative scWGCNA analysis}
\usage{
run.scWGCNA(
  p.cells,
  s.cells,
  idents,
  features,
  is.pseudocell = T,
  min.cells = 10,
  less = F,
  g.names
)
}
\arguments{
\item{p.cells}{Seurat object. The expression data used to run the co-expression analysis. Can be pseudocell or single-cell data but pseudocells are recommended.}

\item{s.cells}{Seurat object. The single cell data, if running on single cell data already, please repeat the argument.}

\item{idents}{Variable. Are certain clusters to be used? Please use group identities and not cell names.}

\item{features}{Variable. The features to be used for the analysis. Default is F, which makes the script calculate variable genes.}

\item{is.pseudocell}{Logical. Is the main data pseudocell data? Default is T}

\item{min.cells}{Numeric. The minimum cells in which genes need to be expressed, to be considered for variable genes calculation. Default is 10}

\item{less}{Logical. In case you are getting too many small modules, use this option as T}

\item{g.names}{Data frame. If you're using gene IDs and no symbols, you might wanna provide a list of gene names for plotting. Two columns: 1= ids present in expression matrix, 2= names to appear in plots. Rownames= same as 1st row}
}
\value{
A list object with the resulting WGCNA data.
}
\description{
This function runs our semi-automatic single-cell WGCNA analysis. It runs in an iterative way. Based on single-cell or pseudocell data.
}
\examples{
# Calculate pseudocells
ps.pbmc_small=calculate.pseudocells(SeuratObject::pbmc_small, dims = 1:10)

# Use pseudocells and single cells to calculate WGCNA
scWGNA.pbmc_small = run.scWGCNA(p.cells = ps.pbmc_small, s.cells = Seurat::pbmc_small)

}
