% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scWGCNA.eigen.R
\name{scW.eigen}
\alias{scW.eigen}
\title{Calculates module eigengenes, average expression, membership and p.values from different sc data.}
\usage{
scW.eigen(modules, seurat)
}
\arguments{
\item{modules}{scWGCNA.data or named vector. An scWGCNA.data object, as calculated by run.scWGCNA().}

\item{seurat}{a Seurat object, which has THE SAME gene names as the ones provided in the modules argument.}
}
\value{
A list with different data
}
\description{
This function calculates module eigengenes, average expression, membership and p.values, from a Seurat object, using as reference a set of gene modules.
}
\examples{
#  We can calculate the eigengenes taking the expression of the single cells
eigengenes.sc = scW.eigen(modules = MmLimbE155.scWGCNA, seurat = my.small_MmLimbE155)

}
