% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scW.p.dendro}
\alias{scW.p.dendro}
\title{Plots the gene / module WGCNA tree, or trees}
\usage{
scW.p.dendro(scWGCNA.data, tree = length(scWGCNA.data$moduletrees.history))
}
\arguments{
\item{scWGCNA.data}{scWGCNA.data. An scWGCNA.data object, as calculated by run.scWGCNA().}

\item{tree}{numeric. Which of all the trees in the iterations should be plotted?}
}
\value{
Plots the last / current gene / module WGCNA dendrogram. If a tree is specified, it returns that tree.
}
\description{
This function will help us plot the gene / module WGCNA tree.
}
\examples{

# A pre-calculated WGCNA data list
class(MmLimbE155.scWGCNA)

# Plot the last, and current WGCNA tree
scW.p.dendro(scWGCNA.data = MmLimbE155.scWGCNA)

# Plot the first WGCNA tree of the iteration
scW.p.dendro(MmLimbE155.scWGCNA, 1)

}
