% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scW.p.expression}
\alias{scW.p.expression}
\title{Plots mean expression of the modules on the single cell data.}
\usage{
scW.p.expression(
  s.cells,
  scWGCNA.data,
  modules = "all",
  reduction = "tsne",
  ncol = 2
)
}
\arguments{
\item{s.cells}{Seurat object. The single cell data used to run scWGCNA.}

\item{scWGCNA.data}{scWGCNA.data. An scWGCNA.data object, as calculated by run.scWGCNA().}

\item{modules}{numeric or "all". Which modules should be plotted? Could be a single module, by number or color, a vector of several modules, or "all" to plot them all in a grid}

\item{reduction}{string. The reduction to use for plots. It must be present in the @reductions slot of sc.data. Default is "tsne"}

\item{ncol}{numeric. If plotting more than 1 module, how many columns should the grid present?}
}
\value{
A ggplot, in case we're plotting only one module, or a grid of plots if plotting several
}
\description{
This function will help us plot the mean expression of the different modules calculated by scWGCNA
}
\examples{
# A pre-analyzed Seurat object, subsampled
my.small_MmLimbE155
MmLimb.sc = my.small_MmLimbE155

# A pre-calculated WGCNA data list
class(MmLimbE155.scWGCNA)
MmLimb.scW = MmLimbE155.scWGCNA

# Plot the expression of all modules
scW.p.expression(s.cells=MmLimb.sc, scWGCNA.data=MmLimb.scW, modules="all", reduction="tsne")

}
