% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scW.p.modulefrac}
\alias{scW.p.modulefrac}
\title{Plots a barplot with the fraction of the expressed and orthologous genes for each module of a scWGCNA comparison.}
\usage{
scW.p.modulefrac(scWGCNA.comp.data)
}
\arguments{
\item{scWGCNA.comp.data}{scWGCNA comparative data as calculated by scWGNA.compare().}
}
\value{
A ggplot barplot
}
\description{
This function will plot a barplot showing what's the fraction of genes in each module that is (if its the case) present in the provided 1-2-1 orthologues table, and the genes that are expressed in all test samples, and therefore used for the comparison.
}
\examples{

# A pre-calculated list scWGCNA comparative data, calculated with scWGCNA.compare
class(MmvGg.comparative)

# Plot the fraction of genes used for the comparison.
scW.p.modulefrac(MmvGg.comparative)
}
