% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{scWGCNA.networks}
\alias{scWGCNA.networks}
\title{Calculates the graphic networks of scWGCNA data}
\usage{
scWGCNA.networks(scWGCNA.data)
}
\arguments{
\item{scWGCNA.data}{scWGCNA.data. An scWGCNA.data object, as calculated by run.scWGCNA().}
}
\value{
An scWGCNA.data object, updated with the slot "networks"
}
\description{
This function calculates the graphic networks of each module in an scWGCNA list object. Returns an updated object.
}
\examples{
# calculate networks
pbmc_small.scWGCNA = scWGCNA.networks(pbmc_small.scWGCNA)

}
