% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scWGCNA.plotexpression}
\alias{scWGCNA.plotexpression}
\title{Plots mean expression of the modules on the single cell data.}
\usage{
scWGCNA.plotexpression(
  s.cells,
  scWGCNA.data,
  modules = "all",
  reduction = "tsne",
  ncol = 2
)
}
\arguments{
\item{s.cells}{Seurat object. The single cell data used to run scWGCNA.}

\item{scWGCNA.data}{scWGCNA.data. An scWGCNA.data object, as calculated by run.scWGCNA().}

\item{modules}{numeric or "all". Which modules should be plotted? Could be a single module, by number or color, a vector of several modules, or "all" to plot them all in a grid}

\item{reduction}{string. The reduction to use for plots. It must be present in the @reductions slot of sc.data. Default is "tsne"}

\item{ncol}{numeric. If plotting more than 1 module, how many columns should the grid present?}
}
\value{
A ggplot, in case we're plotting only one module, or a grid of plots if plotting several
}
\description{
This function will help us plot the mean expression of the different modules calculated by scWGCNA
}
\examples{
# Plot the expression of all modules
scWGCNA.plotexpression(SeuratObject::pbmc_small, pbmc_small.scWGCNA, modules = "all", ncol=3)

}
