% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scWGCNA.plotnetwork}
\alias{scWGCNA.plotnetwork}
\title{Plots a fruchtermanreingold network of the co-expression modules.}
\usage{
scWGCNA.plotnetwork(scWGCNA.data, module = 1, gnames = NULL, ...)
}
\arguments{
\item{scWGCNA.data}{scWGCNA.data. An scWGCNA.data object, as calculated by run.scWGCNA().}

\item{module}{numeric . Which module should be plotted?}

\item{gnames}{Data frame. If you're using gene IDs and no symbols, you might wanna provide a list of gene names for plotting. Two columns: 1= ids present in expression matrix, 2= names to appear in plots. Rownames= same as 1st row}

\item{...}{Parameters passed to GGally::ggnet2()}
}
\value{
A plot showing the relationships between the genes in a module. The layot represents a fruchtermanreingold network. The size of the nodes represent the relative membership of the gene, to the module. The size of the edges represent the relative topology overlap between two genes.
}
\description{
This function will help us plot a fruchtermanreingold network of each of the co-expression modules calculated by scWGCNA.
}
\examples{

# Calculate the networks in the scWCGNA object
MmLimbE155.scWGCNA = scWGCNA.networks(MmLimbE155.scWGCNA)

# Plot one of the modules as network
scWGCNA.plotnetwork(MmLimbE155.scWGCNA, module=1)

# If you're using IDs instead of names, and have a separate gene names data.frame
scWGCNA.plotnetwork(MmLimbE155.scWGCNA, module=1, gnames = my.small_MmLimbE155@misc$gnames)
}
