% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterWGCNA.R
\name{scWGCNA}
\alias{scWGCNA}
\title{scWGCNA}
\usage{
scWGCNA(object, min.module.size = 50, minFraction = 0.25,
  filter_mito_ribo_genes = FALSE, assay_use = NULL, slot_use = NULL,
  merge_similar_modules = FALSE, merge_similarity_threshold = 0.25)
}
\arguments{
\item{object}{Processed scRNAseq object}

\item{min.module.size}{Minimum number of genes needed to form an expression
module. Default: 50.}

\item{minFraction}{When determining genes to keep, what is the minimum
fraction of samples a gene must be detected in before being remove?
Default: 0.25}

\item{filter_mito_ribo_genes}{Should mitochondrial and ribosomal genes be
removed? Default: FALSE}

\item{assay_use}{Instead of using expression data from the obj@data slot,
use data stored in the obj@assay slot}

\item{slot_use}{Assay data slot to use (i.e. "raw.data", "data", or
"scale.data").  Default: 'data'}

\item{merge_similar_modules}{Should similar modules be merged? Default: FALSE}

\item{merge_similarity_threshold}{Dissimilarity (i.e., 1-correlation) cutoff
used to determine if modules should be merged. Default: 0.25}
}
\description{
Perform Weighted Gene Co-expression Network Analysis on scRNAseq
data in a Seurat object
}
\examples{

}
