% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scWGCNA.R
\name{scWGCNA.report}
\alias{scWGCNA.report}
\alias{scWGNA.report}
\title{Runs a semi-automatic, iterative scWGCNA analysis}
\usage{
scWGNA.report(data,sc.data,gene.names,project.name,sp=Mm,cells=F,features=F,reduction="tsne",dir="./",is.pseudocell=T)
}
\arguments{
\item{data}{The expression data to run the analysis in. In a Seurat object. Can be pseudocell or single-cell data. We recommend pseudocells}

\item{sc.data}{The single cell data, if running on single cell data already, please repear the argument}

\item{gene.names}{Translation of the identifiers found in the expression data. Two columns: 1= ids present in expression matrix, 2= names to appear in plots. Can be the same}

\item{project.name}{The name of the project, will be used to save objects and the report}

\item{sp}{The species used for the analysis in string form. Only two letters in the format Sp. Defaults to "Mm"}

\item{cells}{Are certain clusters to be used? Please use identities and not cell names}

\item{features}{The features to be used for the analysis. Default is F, which makes the script calculate variable genes}

\item{reduction}{The reduction to use for plots. It must be present in the @reductions slot of sc.data. Default is "tsne"}

\item{dir}{The directory to save the output to. It defaults to the working directory}

\item{is.pseudocell}{is the main data pseudocell data? Default is T}
}
\value{
Saves a report as well as different objects.
}
\description{
This function runs our semi-automatic single-cell WGCNA analysis. It runs in an iterative way. Based on single-cell or pseudocell data.
}
