% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scWGCNA.compare.R
\name{scWGNA.compare}
\alias{scWGNA.compare}
\title{Runs a comparative scWGCNA analysis}
\usage{
scWGNA.compare(
  scWGCNA.data,
  test.list,
  test.names,
  ortho = NULL,
  ortho.sp = NULL,
  is.pseudocell = T,
  ...
)
}
\arguments{
\item{scWGCNA.data}{Seurat object. The WGCNA data to use as reference for comparative analysis. Output from scWGCNA.report.}

\item{test.list}{List of seurat objects. The samples to test conservation in. If data is calculated on pseudocells, these should also be pseudocells}

\item{test.names}{Vector of strings. Contains the names for each of the samples in test.}

\item{ortho}{Data frame. As many columns as species used only listing 1-to-1 orthologous genes. The species of the reference MUST always be column no. 1}

\item{ortho.sp}{Vector numeric. Indicating to which species in ortho each test sample belongs to.}

\item{is.pseudocell}{Is it?}
}
\value{
No inline output. It saves an html report, as well as an object with the resulting WGCNA comparison indices. They are both named using the project name and the date.
}
\description{
This function runs a WGCNA analysis adapted for single cells. Based on single-cell or pseudocell data.
}
