% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scWGCNA.compare.report.R
\name{scWGNA.compare.report}
\alias{scWGNA.compare.report}
\title{Runs a comparative scWGCNA analysis}
\usage{
scWGNA.compare.report(
  data,
  test,
  test.names,
  project.name,
  ortho,
  ortho.sp,
  groups = F,
  dir = "./",
  interactive = T
)
}
\arguments{
\item{data}{Seurat object. The WGCNA data to use as reference for comparative analysis. Output from scWGCNA.report.}

\item{test}{List of seurat objects. The samples to test conservation in. If data is calculated on pseudocells, these should also be pseudocells}

\item{test.names}{Vector of strings. Contains the names for each of the samples in test.}

\item{project.name}{String. The name of the project, will be used to name the saved objects and the report.}

\item{ortho}{Data frame. As many columns as species used only listing 1-to-1 orthologous genes. The species of the reference MUST always be column no. 1}

\item{ortho.sp}{Vector numeric. Indicating to which species in ortho each test sample belongs to.}

\item{groups}{List of vectors. Indicating if the plots should be grouped. Otherwise, all samples will be plotted together. Suggested for groups of >5 samples. Default is F}

\item{dir}{String. The directory where the output will be saved. It defaults to the working directory.}

\item{interactive}{Logical. Turns off the initial question to continue with the analyses.}
}
\value{
No inline output. It saves an html report, as well as an object with the resulting WGCNA comparison indices. They are both named using the project name and the date.
}
\description{
This function runs a WGCNA analysis adapted for single cells. Based on single-cell or pseudocell data.
}
