% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MmLimbE155.scWGCNA}
\alias{MmLimbE155.scWGCNA}
\title{A pre-computed scWGCNA list calculated from the Mouse limb data.}
\format{
A scWGCNA list object, used as example
}
\source{
This package
}
\usage{
MmLimbE155.scWGCNA
}
\description{
A scWGCNA data list containing the different modules, statistics and calculations
}
\keyword{datasets}
