% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MmvGg.comparative}
\alias{MmvGg.comparative}
\title{Prices of 50,000 round cut diamonds.}
\format{
A Seurat Object, used as example
}
\source{
This package
}
\usage{
MmvGg.comparative
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
