% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my.small_GgLimbHH29}
\alias{my.small_GgLimbHH29}
\title{Prices of 50,000 round cut diamonds.}
\format{
A Seurat object, used as example
}
\source{
This package
}
\usage{
my.small_GgLimbHH29
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
