% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModules.R
\name{plotModules}
\alias{plotModules}
\title{plotModules}
\usage{
plotModules(object, modules, ...)
}
\arguments{
\item{object}{scRNA-seq data object}

\item{modules}{Gene expression modules to plot, as determined by
seuratClusterWGCNA}

\item{...}{Additional arguments to pass to bubbleplot}
}
\value{
gg
}
\description{
Use a bubbleplot to display expression levels and percentage for
genes identified as functioning within a network module.
}
