% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scWGCNA.compare.R
\name{scWGNA.compare}
\alias{scWGNA.compare}
\title{Runs a comparative scWGCNA analysis}
\usage{
scWGNA.compare(
  scWGCNA.data,
  test.list,
  test.names,
  ortho = NULL,
  ortho.sp = NULL,
  is.pseudocell = T,
  ...
)
}
\arguments{
\item{scWGCNA.data}{The WGCNA data to use as reference for comparative analysis. Output from run.scWGCNA}

\item{test.list}{List of Seurat objects. The samples to test conservation in. If data is calculated on pseudocells, these should also be pseudocell objects}

\item{test.names}{Vector of strings. Contains the names for each of the samples in test.}

\item{ortho}{Data frame. If the test objects and the reference data have different feature names, a data frame should be provided. Only listing 1-to-1 orthologous genes. The species of the reference MUST always be column no. 1}

\item{ortho.sp}{Vector numeric. Indicating to which species in ortho each test sample belongs to.}

\item{is.pseudocell}{Logic. Is the analysis based on pseudocells?}

\item{...}{Further parameters, passed to WGCNA::modulePreservation}
}
\value{
It returns a new object, main result of WGCNA::modulePreservation
}
\description{
This function runs a WGCNA analysis adapted for single cells. Based on single-cell or pseudocell data.
}
\examples{
# A pre-calculated WGCNA data list
class(MmLimbE155.scWGCNA)

# A Seurat object with chicken limb cells
my.small_GgLimbHH29

# We calculate pseudocells for this object
GgLimbHH29.ps=calculate.pseudocells(my.small_GgLimbHH29, dims = 1:10)

# Our Seurat objects contain gene names equivalencies in the misc slot. We use them to biuld a table of orthologous genes
my.ortho = merge(my.small_MmLimbE155@misc$gnames,my.small_GgLimbHH29@misc$gnames, by = "Gene.name")
my.ortho=my.ortho[,2:3]

# Comparative tests
MmvGg.comparative = scWGNA.compare(MmLimbE155.scWGCNA, test.list = list(GgLimbHH29.ps), test.names = c("Gg"), ortho = my.ortho, ortho.sp = c(2))
}
